// src/types/purchases.type.ts
import { Product } from './product.type'

export type PurchaseStatus = 1 | 2 | 3 | 4 | 5 | -1

export type PurchaseListStatus = PurchaseStatus | 0

export interface Purchase {
  _id: string
  buy_count: number
  price: number
  price_before_discount: number
  status: PurchaseStatus
  user: string
  product: Product
  createdAt: string
  updatedAt: string
}

export interface ExtendedPurchases extends Purchase {
  disabled: boolean
  checked: boolean
}